pyproject.toml
[project]
name = "hardware-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">3.10"

dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "azure-search-documents==11.4.0b9",
    "black>=25.1.0",
    "fastapi>=0.115.8",
    "fire>=0.7.0",
    "greenlet>=3.1.1",
    "isort>=6.0.0",
    "langchain-community>=0.3.18",
    "langchain-text-splitters>=0.3.6",
    "numpy>=1.26.4",
    "openai>=1.63.2",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pypdf>=5.3.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
